TINKERPOP-1562 Move RemoteAcceptor related stuff to gremlin-core.
